a lot of small corrections. More checks etc.
git-svn-id: https://svn.ic-s.nl/svn/dbmail/trunk/dbmail@1095 7b491191-dbf0-0310-aff6-d879d4d69008

+++ imaputil.h
@@ -60,7 +60,7 @@ int is_textplain(struct list *hdr);
 char *date_sql2imap(const char *sqldate);
 char *date_imap2sql(const char *imapdate);
 
-unsigned stridx(const char *s, char ch);
+size_t stridx(const char *s, char ch);
 int checkchars(const char *s);
 int checktag(const char *s);
 int checkmailboxname(const char *s);
